respond to the case where the class name is written in lower case
