+++ lib/did_you_mean/spell_checkers/name_error_checkers/class_name_checker.rb
@@ -1,10 +1,11 @@
 # frozen-string-literal: true
 require 'delegate'
 require "did_you_mean/spell_checker"
+require "did_you_mean/spell_checkers/name_error_checkers/class_names.rb"
 
 module DidYouMean
   class ClassNameChecker
-    attr_reader :class_name
+    include ClassNames
 
     def initialize(exception)
       @class_name, @receiver, @original_message = exception.name, exception.receiver, exception.original_message
@@ -16,34 +17,5 @@ module DidYouMean
                          .map(&:full_name)
                          .reject {|qualified_name| @original_message.include?(qualified_name) }
     end
-
-    def class_names
-      scopes.flat_map do |scope|
-        scope.constants.map do |c|
-          ClassName.new(c, scope == Object ? "" : "#{scope}::")
-        end
-      end
-    end
-
-    def scopes
-      @scopes ||= @receiver.to_s.split("::").inject([Object]) do |_scopes, scope|
-        _scopes << _scopes.last.const_get(scope)
-      end.uniq
-    end
-
-    class ClassName < SimpleDelegator
-      attr :namespace
-
-      def initialize(name, namespace = '')
-        super(name)
-        @namespace = namespace
-      end
-
-      def full_name
-        self.class.new("#{namespace}#{__getobj__}")
-      end
-    end
-
-    private_constant :ClassName
   end
 end
